Se realiza Instanciada de la clase Observer
