+++ src/main/java/org/example/Main.java
@@ -1,9 +1,9 @@
 package org.example;
 
-import org.example.comprtamiento.Mediator;
+import org.example.comprtamiento.mediator.Mediator;
 import org.example.comprtamiento.Strategy;
+import org.example.comprtamiento.observer.Observer;
 import org.example.creacion.Builder;
-import org.example.creacion.ComidaBuilder;
 import org.example.creacion.Factory;
 import org.example.creacion.Singleton;
 import org.example.estructura.decorator.Decorator;
@@ -16,32 +16,38 @@ public class Main {
 
 
         System.out.println("Hello world!");
+
+        System.out.println();
+        System.out.println("Patron de diseño Builder");
         Builder builder = new Builder();
         builder.ejecutar();
-        System.out.println();
 
+        System.out.println();
+        System.out.println("Patron de diseño Factory");
         Factory factory = new Factory();
         factory.ejecutarCirculo();
         factory.ejecutarTriangulo();
         factory.ejecutarCuadrado();
-        System.out.println();
 
+        System.out.println();
+        System.out.println("Patron de diseño Singleton");
         Singleton singleton = new Singleton();
         singleton.ejecutar();
 
         System.out.println();
-
+        System.out.println("Patron de diseño Strategy");
         Strategy strategy = new Strategy();
         strategy.ejecutarZip();
 
         System.out.println();
-
+        System.out.println("Patron de diseño Decorator");
         Decorator decorator = new Decorator();
         decorator.ejecutarCafe();
         decorator.ejecutarLeche();
 
-        System.out.println();
 
+        System.out.println();
+        System.out.println("Patron de diseño Adapter");
         Adapter adapter = new Adapter();
         adapter.ejecutarMp3();
         adapter.ejecutarWav();
@@ -56,6 +62,11 @@ public class Main {
         Mediator mediator = new Mediator();
         mediator.ejecutarUsuarioNormal();
         mediator.ejecutarChatRoom();
+
+        System.out.println();
+        System.out.println("Patron de diseño Observer");
+        Observer observer = new Observer();
+        observer.ejecutarCondicionActual();
 
     }
 }
