+++ src/main/java/org/example/comprtamiento/observer/DisplayElement.java
@@ -1,0 +1,5 @@
+package org.example.comprtamiento.observer;
+
+interface DisplayElement {
+    void display();
+}
