+++ src/main/java/org/example/comprtamiento/observer/Subject.java
@@ -1,0 +1,9 @@
+package org.example.comprtamiento.observer;
+
+interface Subject {
+    void registerObserver(IObserver o);
+
+    void removeObserver(IObserver o);
+
+    void notifyObservers();
+}
